333337777.17171717

+++ src/main/java/com/edgar/security/repository/TokenRepository.java
@@ -18,5 +18,8 @@ public interface TokenRepository extends JpaRepository<Token, Integer> {
 			select t from Token t inner join User u on t.user.id = u.id
 			where u.id =:userId and (t.expired = false or t.revoked = false)
 			""")
-	List<Token> findAllValidTokensByUser(Integer userId);
+	List<Token> findAllValidTokensByUser(Long userId);
+
+
+	
 }
